@@ -1,17 +1,17 @@
 name: flutter_jsbridge_plugin
-description: use jsbridge to webview in flutter
-version: 0.0.4
+description: use jsbridge to webview in flutter. No reliance on webview_flutter
+version: 1.0.0
 homepage: https://github.com/epoll-j/flutter_jsbridge_plugin
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.22.0"
 
 dependencies:
   flutter:
     sdk: flutter
   # webview_flutter: ^1.0.7
-  sprintf: ^4.0.2
+  sprintf: ^6.0.0
 
 dev_dependencies:
   flutter_test:
