@@ -10,21 +10,21 @@ class Flutterjsbridgeplugin {
     _channel.setMethodCallHandler(_onMethodCall);
   }
 
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
+  static Future<String?> get platformVersion async {
+    final String? version = await _channel.invokeMethod('getPlatformVersion');
     return version;
   }
 
   static Future<String> get scrpit async {
-    final String initScript = await _channel.invokeMethod("getScript");
+    final String? initScript = await _channel.invokeMethod("getScript");
     return "javascript:$initScript";
   }
 
-  Future<bool> isMainThread() async {
+  Future<bool?> isMainThread() async {
     return await _channel.invokeMethod("isMainThread");
   }
 
-  Future<String> handlerReturnData(String url) async {
+  Future<String?> handlerReturnData(String url) async {
     return await _channel.invokeMethod("handlerReturnData", url);
   }
 
